+++ services/has-many-getter.js
@@ -4,10 +4,12 @@ var P = require('bluebird');
 var SearchBuilder = require('./search-builder');
 var Interface = require('forest-express');
 var utils = require('../utils/schema');
+var RecordsDecorator = require('../utils/records-decorator');
 
 function HasManyGetter(model, association, opts, params) {
   var OBJECTID_REGEXP = /^[0-9a-fA-F]{24}$/;
   var schema = Interface.Schemas.schemas[utils.getModelName(association)];
+  var searchBuilder = new SearchBuilder(association, opts, params);
   var count = 0;
 
   function hasPagination() {
@@ -73,8 +75,7 @@ function HasManyGetter(model, association, opts, params) {
         };
 
         if (params.search) {
-          var conditionsSearch = new SearchBuilder(association, opts, params)
-            .getConditions();
+          var conditionsSearch = searchBuilder.getConditions();
           conditions.$and.push(conditionsSearch);
         }
 
@@ -111,7 +112,20 @@ function HasManyGetter(model, association, opts, params) {
   this.perform = function () {
     return getRecords()
       .then(function (records) {
-        return [count, records];
+        var decorators = null;
+
+        if (params.search) {
+          var decoratorsSearch = RecordsDecorator.decorateForSearch(
+            getRecords(),
+            searchBuilder.getFieldsSearched(),
+            params.search
+          );
+          if (!_.isEmpty(decoratorsSearch)) {
+            decorators = decoratorsSearch;
+          }
+        }
+
+        return [records, count, decorators];
       });
   };
 }