+++ CHANGELOG.md
@@ -2,7 +2,10 @@
 
 ## [Unreleased]
 ### Added
-- Search & Technical - Add tests for search on selected field & switch linter to eslint.
+- Search - Add tests for search on selected field.
+
+### Changed
+- Technical - Switch the linter to eslint.
 
 ## RELEASE 2.8.3 - 2018-04-25
 ### Fixed