+++ CHANGELOG.md
@@ -4,12 +4,12 @@
 ### Added
 - Search - Add tests for search on selected field.
 
-### Fixed
-- Technical - Use local packages for npm scripts.
-
 ### Changed
 - Technical - Switch the linter to eslint.
 - Technical - Save exact dependencies version.
+
+### Fixed
+- Technical - Use local packages for npm scripts.
 
 ## RELEASE 2.8.4 - 2018-04-30
 ### Fixed